Struct rtlsdr_mt::Controller

pub struct Controller(_);

Controls hardware parameters.

Methods

impl Controller

pub fn sample_rate(&self) -> u32

Get the current sample rate (megasamples/sec).

pub fn set_sample_rate(&mut self, rate: u32) -> Result<()>

Set the sample rate (megasamples/sec).

pub fn center_freq(&self) -> u32

Get the current center frequency (Hz).

pub fn set_center_freq(&mut self, freq: u32) -> Result<()>

Set the center frequency (Hz).

pub fn ppm(&self) -> i32

Get the current frequency correction (ppm).

pub fn set_ppm(&mut self, ppm: i32) -> Result<()>

Set the frequency correction (ppm).

pub fn enable_agc(&mut self) -> Result<()>

Enable the hardware AGC.

Note that this also disables manual tuner gain.

pub fn disable_agc(&mut self) -> Result<()>

Disable the hardware AGC.

Note that this also enables manual tuner gain.

pub fn tuner_gains<'a>(&self, gains: &'a mut TunerGains) -> &'a [i32]

Get the list of valid tuner gain values.

Each value represents a dB gain with the decimal place shifted right. For example, the value 496 represents 49.6dB.

pub fn tuner_gain(&self) -> i32

Get the current tuner gain in the same format as that returned by tuner_gains().

pub fn set_tuner_gain(&mut self, gain: i32) -> Result<()>

Set the tuner gain in the same format as that returned by tuner_gains().

Note that this also disables the hardware AGC.

pub fn cancel_async_read(&mut self)

Cancel an asynchronous read if one is running.