Struct rtlsdr_mt::Reader

pub struct Reader(_);

Reads I/Q samples.

Methods

impl Reader

pub fn read_async<F>(&mut self, bufs: u32, len: u32, cb: F) -> Result<()> where
F: FnMut(&[u8]),

Begin reading I/Q samples, buffering into the given number of chunks, with each chunk holding the given number of bytes. The given callback is called whenever new samples are available, receiving a chunk at a time.

This function blocks until the read is cancelled or otherwise terminated. Hardware parameters can be changed in a separate thread while this function is running.